+++ Artisan/ContextMenus/CraftingListContextMenu.cs
@@ -49,9 +49,9 @@ internal static class CraftingListContextMenu
             if (RetainerInfo.GetRetainerItemCount(itemId) > 0 && RetainerInfo.GetReachableRetainerBell() != null)
             {
                 int amountToGet = 1;
-                if (LuminaSheets.RecipeSheet[(uint)Endurance.RecipeID].ItemResult.Row != itemId)
+                if (LuminaSheets.RecipeSheet[Endurance.RecipeID].ItemResult.Row != itemId)
                 {
-                    amountToGet = LuminaSheets.RecipeSheet[(uint)Endurance.RecipeID].UnkData5.First(y => y.ItemIngredient == itemId).AmountIngredient;
+                    amountToGet = LuminaSheets.RecipeSheet[Endurance.RecipeID].UnkData5.First(y => y.ItemIngredient == itemId).AmountIngredient;
                 }
 
                 args.AddCustomItem(new GameObjectContextMenuItem(new Dalamud.Game.Text.SeStringHandling.SeString(new UIForegroundPayload(706), new TextPayload($"{SeIconChar.BoxedLetterA.ToIconString()} "), UIForegroundPayload.UIForegroundOff, new TextPayload("Withdraw from Retainer")), _ => RetainerInfo.RestockFromRetainers(itemId, amountToGet)));
